server/requirements.txt
fastapi[standard]
pymupdf
python-dotenv
jinja2


# Embeddings
langchain
langchain-google-genai
langchain-text-splitters
langchain-core
numpy